good doc on blockship

+++ ft_transcendence_db/src/entity/blockships/blockship.gateway.ts
@@ -18,14 +18,22 @@ export class BlockshipGateway {
     constructor(private readonly blockshipService: BlockshipService, private readonly sessionService: SessionService) {
     }
 
+    /**
+     * remove blockship
+     * 
+     * @param body sessionCookie of user making request and user blocked
+     * @returns 
+     */
     @SubscribeMessage('removeBlockship')
-    async refuseBlockship(@MessageBody() body) 
+    async refuseBlockship(@MessageBody() body: {sessionCookie: string, userblocked_id: number}) 
     {
       if (await this.sessionService.getIsSessionExpired(body.sessionCookie))
       {
         return ('not connected');
       }
       const user = await this.sessionService.getUser(body.sessionCookie);
+      if (!user)
+        return (null);
       await this.blockshipService.deleteBlockship(user.id, body.userblocked_id);
       this.server.emit('deleteBlockship', {sessionCookie: body.sessionCookie});
     }
